fix: some minor fixes
